Big refactor (remove extra & rework according issue)
