--- src/views/event-item/create-trip-event-item-edit/create-event-item-collapsible-btn.js
@@ -1,7 +1,0 @@
-export const createEventItemCollapsibleBtn = () => (
-  `
-  <button class="event__rollup-btn" type="button">
-    <span class="visually-hidden">Open event</span>
-  </button>
-  `
-);
