--- src/views/event-item/create-trip-event-item-edit/create-trip-event-destination-description.js
@@ -1,5 +1,0 @@
-export const createTripEventDestinationDescription = () => (
-  `
-  <p class="event__destination-description">Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.</p>
-  `
-);
