--- src/views/event-item/create-trip-event-item-edit/create-trip-event-destination-photos.js
@@ -1,13 +1,0 @@
-export const createTripEventDestinationPhotos = () => (
-  `
-  <div class="event__photos-container">
-    <div class="event__photos-tape">
-      <img class="event__photo" src="img/photos/1.jpg" alt="Event photo">
-      <img class="event__photo" src="img/photos/2.jpg" alt="Event photo">
-      <img class="event__photo" src="img/photos/3.jpg" alt="Event photo">
-      <img class="event__photo" src="img/photos/4.jpg" alt="Event photo">
-      <img class="event__photo" src="img/photos/5.jpg" alt="Event photo">
-    </div>
-  </div>
-  `
-);
