--- src/views/event-item/create-trip-event-item-edit/create-trip-event-destination-wrapper.js
@@ -1,7 +1,0 @@
-export const createTripEventDestinationWrapper = () => (
-  `
-  <section class="event__section  event__section--destination">
-    <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-  </section>
-  `
-);
