--- src/views/event-item/create-trip-event-item-edit/create-trip-event-item-edit-wrapper.js
@@ -1,8 +1,0 @@
-export const createTripEventItemEditWrapper = () => (
-  `
-  <li class="trip-events__item">
-    <form class="event event--edit" action="#" method="post">
-    </form>
-  </li>
-  `
-);
