+++ src/views/header/trip-filters-header.js
@@ -1,7 +1,7 @@
 export const createTripFilterHeaderTemplate = () => (
   `
-<div class="trip-controls__filters">
-  <h2 class="visually-hidden">Filter events</h2>
+  <div class="trip-controls__filters">
+    <h2 class="visually-hidden">Filter events</h2>
     <form class="trip-filters" action="#" method="get">
       <div class="trip-filters__filter">
         <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
