+++ src/views/header/trip-tabs-header.js
@@ -1,4 +1,4 @@
-export const createTripTabsHeader = () => (
+export const createTripTabsHeaderTemplate = () => (
   `<div class="trip-controls__navigation">
     <h2 class="visually-hidden">Switch trip view</h2>
       <nav class="trip-controls__trip-tabs  trip-tabs">
