+++ src/views/trip-event-item.js
