+++ src/views/trip-events-list-wrapper.js
